style(css): add responsive header, footer, post date
This commit adds a change in the width of the post date, when the
window width is 380 pixels, and presses the header and footer to the
edge of the page in case there is not enough content in main to fill
the page. This is done for a more comfortable use of the site.