style(css, scripts): formatting
This commit adds formatting to CSS and JavaScript files. In the
JavaScript file, double brackets in the array of posts have been
replaced with single ones, and one apostrophe has been replaced with
a mnemonic. In the CSS file, unnecessary spaces have been removed.
This is necessary to improve the readability of the code and its
appearance.